Addded constraint variable for 'ALL.pdf'.
The constraint variable __all_same_page can now be set in the constraint file.
If set to True then all the forms will print consecutively on the same page if
possible. If set to False then each form in 'ALL.pdf' will start on a new page.
If left unspecified, default is False.