Add '__header_box' option for the const files.
The '__header_box' option allows the user to further customize the heading of
the forms. The '__header_box' option values can be either box, form_name or None. If the option
box is selected, a box will be printed in the upper right hand corner of every
page with clinician name, date, and study id inside. If the option form_name is
selected, the form name will be printed on the top of each page. If None is
selected then nothing will be printed in the top right corner of the page.