Add feature '__print_selected_only' for config file.
The user can now specify '__print_selected_only = True' in the config file.
For multi choice questions that have constraints in the config file, this will
cause only the valid/acceptable choices to print instead of printing all of
them. If __print_selected_only is not specified, it will default to False,
causing all choices to print.