Add ability to specify regex to supress printing elements.
Users can now specify __no_print_regex in the config file followed by a list of
regular expressions. Any elements that have field names that match any of the
regular expression will not be printed on the forms. Elements nested under the
field_name not printed is also not printe.